Return anonymous functions instead of named functions for benchmarkable entry points
This change fixes method overwrite warnings when generating benchmarkable entry points in loop-like contexts on Julia v0.5. This changes the benchmarkable macro API so that the caller can simply bind the return value to a variable rather than interpolating variable names into the macro context. Temporary functions are wrapped in a let block so that their bindings are freed upon exiting the scope, which fixes overwrite warnings for these functions as well.